+++ qiskit_ibm_runtime/utils/__init__.py
@@ -19,12 +19,8 @@ from .converters import (
 from .utils import (
     is_crn,
     default_runtime_url_resolver,
-    resolve_crn,
     are_circuits_dynamic,
     cname_from_crn,
-    get_iam_api_url,
-    get_global_search_api_url,
-    get_global_catalog_api_url,
 )
 from .validations import (
     validate_estimator_pubs,
